Split walk-in onboarding stages and controls
